Categorize ACS variables about children living in poverty and parents' birth origins

These data come from table B05010, "RATIO OF INCOME TO POVERTY LEVEL IN THE PAST 12 MONTHS BY NATIVITY OF CHILDREN UNDER 18 YEARS IN FAMILIES AND SUBFAMILIES BY LIVING ARRANGEMENTS AND NATIVITY OF PARENTS."

Usage

CHILDREN_IN_POVERTY_METADATA

Format

CHILDREN_IN_POVERTY_METADATA

A data frame with 15 rows and 8 columns

Group

The table, always "B05010"

Variable

The full variable name, e.g. "B05010_001E"

Index

The row of this variable in the table

Least Poverty Ratio

The lowest ratio of income to poverty level in this tier

Greatest Poverty Ratio

The highest ratio of income to poverty level in this tier

Standard of Living

One of "Unsustainable," or "Mixed"

Native-born Parents

How many of the parents in the household were born in the USA.

Foreign-born Parents

How many of the parents in the household were not born in the USA.

Source

https://api.census.gov/data/2022/acs/acs1/groups/B05010.html