For selected tables, an alphabetic suffix follows to indicate that a table is repeated for the nine major race and Hispanic or Latino groups:

For selected tables, an alphabetic suffix follows to indicate that a table is repeated for the nine major race and Hispanic or Latino groups:

Usage

RACE_OR_ETHNIC_GROUP_TYPES

Format

RACE_OR_ETHNIC_GROUP_TYPES

A data frame with 9 rows and 2 columns.

Code

<chr> A one-letter suffix

Race or Ethnic Group

<chr> A racial or ethnic identity such as "Black Alone" or "Hispanic or Latino".

Source

https://www.census.gov/programs-surveys/acs/data/data-tables/table-ids-explained.html