Pull data from the 5-year American Community Survey

The default years span the entire offering of 5-year ACS data. Your variables will probably not all be available for every year.

Usage

fetch_acs5_tract_data(variables, state, county, years = 2009:2021L)

Arguments

variables

a character vector of variables like "B03002_001E"

state

the two-digit FIPS code for the state, like 55 for Wisconsin

county

the three-digit FIPS code for the county, like 101 for Racine

years,

optional, the years to fetch data for. Defaults to 2009-2021

Value

a tibble of year, vintage, tract, group, index, variable, and value