Get decennial census estimates of per-block populations by race/ethnicity

This function retrieves data from tables that splits people into race and Hispanic/Latin\@ ethnicity.

Usage

fetch_decennial_block_pops_by_race(state, county)

Arguments

state

two-digit FIPS code for the state, e.g. 55L for Wisconsin

county

three-digit FIPS code for the county within the state, e.g. 101L for Racine

Value

a tibble with four columns:

Vintage

The decennial year, currently 2000, 2010, or 2020

Race/Ethnicity

OMB text descriptions of Race/Ethnicities

GEOID

The full FIPS code for each census block

Population

The number of people in that block in that year of that race/ethnicity