Determine the relative contribution of different demographic groups to the populations of a set of areas.

Determine the relative contribution of different demographic groups to the populations of a set of areas.

Usage

margins_by_demography_and_place(
  block_data,
  ...,
  demography_field = "Race/Ethnicity",
  place_field = "Place",
  weight_field = "Population"
)

Arguments

block_data

a tibble of census data that connects blocks to enclosing geographic areas.

...

<dynamic-dots> grouping fields in the data besides demography and place

demography_field

optional, the field that identifies peoples' demographic group. Defaults to "Race/Ethnicity"

place_field

optional, the field that denotes your custom geographic areas. Defaults to "Place"

weight_field

optional, the field that denotes counts. Defaults to "Population"

Value

a tibble with proportions for estimating per-demographic values from per-tract data