The fundamental demographics of age and sex from table B01001

The age splits are more detailed in the whole-population table than they are in the tables that pertain to specific racial identities. A future project is to add a second part to the table that maps from those tables' rows to age and sex.

Usage

GLOSSARY_OF_AGE_AND_SEX

Format

GLOSSARY_OF_AGE_AND_SEX

A data frame with 49 rows and 6 columns.

group

<chr> Always "B01001"

index

<int> The row in the source table that this GLOSSARY row describes.

variable

<chr> The full variable name, e.g. "B01001_001E"

Sex

<chr> One of "All," "Male," or "Female"

Lower Age

<int> inclusive, in [-999, 85]

Upper Age

<int> inclusive, in [4, 999]

Source

https://api.census.gov/data/2022/acs/acs1/groups/B01001.html