ACS data on sex, age, and labor force status

ACS data on sex, age, and labor force status

Usage

GLOSSARY_OF_LABOR_FORCE_PARTICIPATION

Format

GLOSSARY_OF_LABOR_FORCE_PARTICIPATION

A data frame with 172 rows and 10 columns

Dataset

<chr> "ACS1" or "ACS5"

Group

<chr> Always "B23001"

Index

<int> the row number of the variable

Variable

<chr> the full variable

Sex

<chr> "Male" or "Female"

Lower Age

<int> between 16 and 75, inclusive

Upper Age

<int> between 19 and 999, inclusive

Labor Force

<lgl>

Civilian

<lgl>

Employed"

<lgl>